Use docutils as a Python library to generate man pages
This greatly simplifies the setup.py code, albeit
at the cost that should setuptools not be
available or outdated so that the docutils egg
cannot be downloaded and docutils not be installed
for the Python version used, no man pages will be
generated.

Now this is a warning only instead of an error.