Split tests out to separate package
With the goal of skonfig being to be easily installable without (deprecated)
setuptools or library managers such as pip, it makes more sense to have unit
tests split from the rest of the code base.

This is to have a package which contains only the code necessary to run the
application.  This (single) package can then also be easily distributed and
downloaded by users.

Since most users will not want to run the test suite, we don't have to have
them all install the test suite.