Replace glob.glob1() with custom listdir function
We have to replace the glob.glob1() function because Python will remove it in
the 3.15 release.
Since the use of glob1() in skonfig is limited to listing directory contents, I
wrote a replacement skonfig.util.(i)listdir() which is also prepared to exclude
unwanted file (like save files *~) and for recursive listing of directory
contents (currently not used but still useful to have).

    DeprecationWarning: glob.glob1 is deprecated and will be removed in Python 3.15. Use glob.glob and pass a directory to its root_dir argument instead.